+++ backend/src/main/java/undecided/shared/common/annotation/CanIgnoreReturnValue.java
@@ -1,4 +1,4 @@
-package undecided.erp.common.annotation;
+package undecided.shared.common.annotation;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -8,12 +8,10 @@ import java.lang.annotation.Target;
 
 /**
  * このアノテーションは、メソッドまたはコンストラクタの戻り値を無視できることを示すために使用できます。
- * <p>
- * メソッド、コンストラクタ、型に適用できます。
+ *
+ * <p>メソッド、コンストラクタ、型に適用できます。
  */
 @Documented
 @Target({ElementType.METHOD, ElementType.CONSTRUCTOR, ElementType.TYPE})
 @Retention(RetentionPolicy.CLASS)
-public @interface CanIgnoreReturnValue {
-
-}
+public @interface CanIgnoreReturnValue {}
