+++ backend/src/main/java/undecided/shared/common/annotation/DoNotCall.java
@@ -1,13 +1,11 @@
-package undecided.erp.common.annotation;
+package undecided.shared.common.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-/**
- * 呼ぶべきではないメソッドに付与する
- */
+/** 呼ぶべきではないメソッドに付与する */
 @Retention(RetentionPolicy.CLASS)
 @Target({ElementType.METHOD})
 public @interface DoNotCall {
@@ -18,5 +16,4 @@ public @interface DoNotCall {
    * @return メソッドに関する説明や意図した使用制限事項
    */
   String value() default "";
-
 }
