+++ backend/src/test/java/undecided/shared/common/builder/AbstractBuilderTest.java
@@ -1,4 +1,4 @@
-package undecided.erp.common.builder;
+package undecided.shared.common.builder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -6,14 +6,10 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
-/**
- * AbstractBuilderクラスのためのテストクラスです。
- */
+/** AbstractBuilderクラスのためのテストクラスです。 */
 class AbstractBuilderTest {
 
-  /**
-   * このクラスはConcreteClassとConcreteRecordビルダーのbuild()メソッドのテストを含んでいます。
-   */
+  /** このクラスはConcreteClassとConcreteRecordビルダーのbuild()メソッドのテストを含んでいます。 */
   @Nested
   @DisplayName("buildメソッドテスト")
   class BuildTest {
@@ -21,10 +17,7 @@ class AbstractBuilderTest {
     @Test
     @DisplayName("ConcreteClassがBuilderから正しく作成される")
     void shouldCreateConcreteClassFromBuilder() {
-      ConcreteClass actualClass = ConcreteClass.builder()
-          .withId(1)
-          .withName("name")
-          .build();
+      ConcreteClass actualClass = ConcreteClass.builder().withId(1).withName("name").build();
       assertThat(actualClass).isNotNull();
       System.out.println(actualClass);
     }
@@ -32,19 +25,13 @@ class AbstractBuilderTest {
     @Test
     @DisplayName("ConcreteRecordがBuilderから正しく作成される")
     void shouldCreateConcreteRecordFromBuilder() {
-      ConcreteRecord actualClass = ConcreteRecord.builder()
-          .withId(1)
-          .withName("name")
-          .build();
+      ConcreteRecord actualClass = ConcreteRecord.builder().withId(1).withName("name").build();
       assertThat(actualClass).isNotNull();
       System.out.println(actualClass);
     }
-
   }
 
-  /**
-   * ApplyTest クラスは ConcreteClass と ConcreteRecord の apply() メソッドをテストする責任があります。
-   */
+  /** ApplyTest クラスは ConcreteClass と ConcreteRecord の apply() メソッドをテストする責任があります。 */
   @Nested
   @DisplayName("applyメソッドテスト")
   class ApplyTest {
@@ -53,12 +40,13 @@ class AbstractBuilderTest {
     @DisplayName("ConcreteClassがApplyメソッドを用いて正しく作成される")
     void shouldCreateConcreteClassUsingApplyMethod() {
       ConcreteClass source = new ConcreteClass(2, "name2");
-      ConcreteClass actualClass = ConcreteClass.builder()
-          .withId(null)
-          .withName(null)
-          .withId(1)
-          .withName("name")
-          .apply(source);
+      ConcreteClass actualClass =
+          ConcreteClass.builder()
+              .withId(null)
+              .withName(null)
+              .withId(1)
+              .withName("name")
+              .apply(source);
       assertThat(actualClass).isNotNull();
       System.out.println(actualClass);
     }
@@ -67,15 +55,15 @@ class AbstractBuilderTest {
     @DisplayName("ConcreteRecordがApplyメソッドを用いて正しく作成される")
     void shouldCreateConcreteRecordUsingApplyMethod() {
       ConcreteRecord source = new ConcreteRecord(2, "name2");
-      ConcreteRecord actualClass = ConcreteRecord.builder()
-          .withId(null)
-          .withName(null)
-          .withId(1)
-          .withName("name")
-          .apply(source);
+      ConcreteRecord actualClass =
+          ConcreteRecord.builder()
+              .withId(null)
+              .withName(null)
+              .withId(1)
+              .withName("name")
+              .apply(source);
       assertThat(actualClass).isNotNull();
       System.out.println(actualClass);
     }
-
   }
 }
