+++ backend/src/test/java/undecided/shared/common/builder/ConcreteRecord.java
@@ -1,8 +1,6 @@
-package undecided.erp.common.builder;
+package undecided.shared.common.builder;
 
-public record ConcreteRecord(Integer id,
-                             String name) {
-
+public record ConcreteRecord(Integer id, String name) {
 
   /**
    * ConcreteBuilderの新しいインスタンスを返します。
@@ -13,8 +11,8 @@ public record ConcreteRecord(Integer id,
     return new ConcreteRecordBuilder();
   }
 
-  public static class ConcreteRecordBuilder extends
-      AbstractBuilder<ConcreteRecord, ConcreteRecordBuilder> {
+  public static class ConcreteRecordBuilder
+      extends AbstractBuilder<ConcreteRecord, ConcreteRecordBuilder> {
 
     private Integer id;
     private String name;
@@ -23,13 +21,11 @@ public record ConcreteRecord(Integer id,
     protected void apply(ConcreteRecord vo, ConcreteRecordBuilder builder) {
       builder.withId(vo.id());
       builder.withName(vo.name());
-
     }
 
     public ConcreteRecordBuilder withId(Integer id) {
       configurators.add(builder -> builder.id = id);
       return getThis();
-
     }
 
     public ConcreteRecordBuilder withName(String name) {
@@ -52,5 +48,4 @@ public record ConcreteRecord(Integer id,
       return new ConcreteRecordBuilder();
     }
   }
-
 }
